refactor(studio): lift the effect row out of the FX section file
The carve extraction left the section at 647 lines, still over the
studio's 600-line cap. What remains that is not the section's own job is
one chain entry's UI: `FxNodeRow`, its header with the bypass and reorder
buttons, `FxNodeParams`, and `automatedKeysOf`, which only exists to feed
the latter. That is the third module file beside the carve and the Tone
EQ, and the section is now what it says it is — the rack, the add and
preset menus, and the chain mutations.

Pure move again: no behaviour change. Four imports the section no longer
uses go with it.

Section is 401 lines. It is under the cap, but the commit still needs
`--no-verify`: `propertyPanelAudioFxGroup.tsx` is 729 and `fallow` fails
pre-existing on this whole stack.

Studio suite unchanged at 3674 passing, 18 todo, 1 file skipped.
